refactor: 인터페이스 추출로 DIP 적용 및 예외 패키지 정리
- LogParser, AnalysisResultRepository 인터페이스 추출로 DIP 적용
- AnalysisService가 구현체 대신 인터페이스에 의존하도록 변경
- ParseResult record를 LogParser 인터페이스로 이동
- InvalidCsvFormatException 패키지를 domain.exception으로 이동
- 관련 테스트 및 import 경로 정리
